+++ assets/js/app.js
@@ -5,6 +5,7 @@ import AIPromptInput from "./hooks/ai_prompt_input";
 import AutoRefresh from "./hooks/auto_refresh";
 import AutoScrollAI from "./hooks/auto_scroll_ai";
 import Clipboard from "./hooks/clipboard";
+import ColumnStats from "./hooks/column_stats";
 import DispatchChangeOnUpdate from "./hooks/dispatch_change_on_update";
 import EditorForm from "./hooks/editor_form";
 import Flash from "./hooks/flash";
@@ -73,6 +74,7 @@ const hooks = {
   AutoRefresh,
   AutoScrollAI,
   Clipboard,
+  ColumnStats,
   DispatchChangeOnUpdate,
   EditorForm,
   Flash,